example: add summary module, backend, and frontend
Summary:
The example plugin becomes a bit more like an actual TensorBoard plugin:
you can write data to it, and it can render that data.

Test Plan:
Smoke test expanded, and the UI looks like this:

![Screenshot of example plugin table view with demo data]

[1]: https://user-images.githubusercontent.com/4317806/59955944-e8247800-9441-11e9-9cd3-bf978e282717.png

wchargin-source: d37a71aff17510abdd25e2c89073b8b9f8376603
wchargin-branch: example-summary

+++ tensorboard/plugins/example/smoke_test.sh
@@ -44,18 +44,51 @@ pip install -qU wheel 'setuptools>=36.2.0'
 (cd ./example-plugin && python setup.py bdist_wheel)
 [ -f ./example-plugin/dist/*.whl ]  # just one wheel
 
-# This plugin doesn't require TensorFlow, so we don't install it.
 py_major_version="$(python -c 'import sys; print(sys.version_info[0])')"
+pip install tf-nightly-2.0-preview  # TODO(@wchargin): Other versions, too?
+pip uninstall -y tensorboard tb-nightly  # drop conflicting packages
 pip install ./tensorboard-wheels/*py"${py_major_version}"*.whl
 pip install ./example-plugin/dist/*.whl
 
+python -m tensorboard_plugin_example.demo
+
 # Test tensorboard + tensorboard_plugin_example integration.
 mkfifo pipe
-tensorboard --port=0 --logdir=smokedir 2>pipe &
+tensorboard --port=0 --logdir=. 2>pipe &
 perl -ne 'print STDERR;/http:.*:(\d+)/ and print $1.v10 and exit 0' <pipe >port
-curl -fs "http://localhost:$(cat port)/data/plugins_listing" >plugins_listing
-grep -F '"example":' plugins_listing
-grep -F '"/data/plugin/example/index.js"' plugins_listing
-curl -fs "http://localhost:$(cat port)/data/plugin/example/index.js" >index.js
+port="$(cat port)"
+curl -fs "http://localhost:${port}/data/plugin/example/index.js" >index.js
 diff -u example-plugin/tensorboard_plugin_example/static/index.js index.js
+
+# Wait for data to be read so that the plugin becomes active.
+for attempt in {1..5}; do
+  sleep 1
+  curl -fs "http://localhost:${port}/data/plugins_listing" >plugins_listing
+  status="$(python3 -c '
+import sys
+import json
+with open("plugins_listing") as infile:
+  plugins = json.load(infile)
+if "example" in plugins:
+  print("active" if plugins["example"]["enabled"] else "inactive")
+else:
+  print("missing")
+  sys.stderr.write("%s\n" % (plugins,))
+  ')"
+  if [ "${status}" = missing ]; then
+    printf >&2 'example plugin not picked up by TensorBoard!\n'
+    exit 1
+  fi
+done
+if [ "${status}" != active ]; then
+  printf >&2 'plugin data still not loaded; cannot verify that demo worked\n'
+  exit 1
+fi
+grep -qF '"/data/plugin/example/index.js"' plugins_listing
+
+curl -fs "http://localhost:${port}/data/plugin/example/tags" >tags
+<<EOF tr -d '\n' | diff -u - tags
+{"demo_logs": {"guestbook": {"description": "Sign your name!"}, "more_names": {"description": ""}}}
+EOF
+
 kill $!
